requirements.txt
discord.py
flask
gunicorn
python-dotenv